restructure zeus component ui files
