+++ addons/zeus/UI.hpp
@@ -1,180 +1,4 @@
-class RscText;
-class RscButtonMenu;
-class RscListbox;
-class RscCheckBox;
-class RscControlsGroup;
-class RscControlsGroupNoScrollbars;
-class RscDisplayAttributes {
-    class Controls {
-        class Background;
-        class Title;
-        class Content: RscControlsGroup {
-            class controls;
-        };
-        class ButtonOK;
-        class ButtonCancel;
-    };
-};
-
-
-//supply drop ==================================================================
-class GVAR(moduleSupplyDrop) {
-    idd = 3646235;
-    movingEnable = true;
-
-    class controlsBackground {
-        class Title: RscText {
-        	text = "Supply drop";
-        	x = 0.360781 * safezoneW + safezoneX;
-        	y = 0.214 * safezoneH + safezoneY;
-        	w = 0.262969 * safezoneW;
-        	h = 0.022 * safezoneH;
-        	colorBackground[] = {-1,-1,-1,0.7};
-        };
-
-        class Background: RscText {
-        	x = 0.360781 * safezoneW + safezoneX;
-        	y = 0.236 * safezoneH + safezoneY;
-        	w = 0.262969 * safezoneW;
-        	h = 0.528 * safezoneH;
-        	colorBackground[] = {-1,-1,-1,0.4};
-        };
-    };
-
-    class controls {
-        class PlaneList: RscListbox {
-        	idc = 1500;
-        	x = 0.365937 * safezoneW + safezoneX;
-        	y = 0.247 * safezoneH + safezoneY;
-        	w = 0.252656 * safezoneW;
-        	h = 0.506 * safezoneH;
-        };
-
-        class ConfirmButton: RscButtonMenu {
-        	idc = 2400;
-        	text = "Anfordern";
-        	x = 0.561615 * safezoneW + safezoneX;
-        	y = 0.76713 * safezoneH + safezoneY;
-        	w = 0.0626562 * safezoneW;
-        	h = 0.0252408 * safezoneH;
-        };
-
-        class AbortButton: RscButtonMenu {
-        	idc = 2401;
-        	text = "Abbrechen";
-        	x = 0.497395 * safezoneW + safezoneX;
-        	y = 0.766667 * safezoneH + safezoneY;
-        	w = 0.0626562 * safezoneW;
-        	h = 0.0252408 * safezoneH;
-        };
-    };
-};
-
-
-//diagnostics ==================================================================
-class GVAR(RscModuleDiagnostics): RscDisplayAttributes {
-    onLoad = QUOTE([ARR_3('onLoad', _this, QUOTE(QGVAR(RscModuleDiagnostics)))] call ace_zeus_fnc_zeusAttributes);
-    onUnload = QUOTE([ARR_3('onUnload', _this, QUOTE(QGVAR(RscModuleDiagnostics)))] call ace_zeus_fnc_zeusAttributes);
-    class Controls: Controls {
-        class Background: Background {};
-        class Title: Title {};
-        class Content: Content {
-            class Controls {
-                class globalSetSkill: RscControlsGroupNoScrollbars {
-                    onSetFocus = QUOTE(_this call FUNC(ui_moduleDiagnostics));
-                    idc = 27000;
-                    x = 0;
-                    y = 0;
-                    w = W_PART(26);
-                    h = H_PART(6.5);
-                    class controls {
-                        class Title1: RscText {
-                            idc = -1;
-                            /* text = CSTRING(ASD); */
-                            text = "Player FPS";
-                            /* toolTip = CSTRING(ASD); */
-                            toolTip = "Show player FPS";
-                            x = 0;
-                            y = 0;
-                            w = W_PART(10);
-                            h = H_PART(1);
-                            colorBackground[] = {0,0,0,0.5};
-                        };
-                        class PlayerFPS: RscCheckBox {
-                            idc = 27001;
-                            x = W_PART(10.1);
-                            y = 0;
-                            w = W_PART(1);
-                            h = H_PART(1);
-                        };
-                        class Title2: Title1 {
-                            idc = -1;
-                            text = "Player medical status";
-                            tooltip = "Show medical status of players.";
-                            y = H_PART(1.1);
-                        };
-                        class MedicalStatus: PlayerFPS {
-                            idc = 27002;
-                            y = H_PART(1.1);
-                        };
-                        class Title3: Title1 {
-                            idc = -1;
-                            text = "Object locality";
-                            tooltip = "Show owner of objects and units.";
-                            y = H_PART(2.2);
-                        };
-                        class ObjectLocality: PlayerFPS {
-                            idc = 27003;
-                            y = H_PART(2.2);
-                        };
-                        class Title4: Title1 {
-                            idc = -1;
-                            text = "AI status";
-                            tooltip = "Show combat mode, fleeing, etc.";
-                            y = H_PART(3.3);
-                        };
-                        class AIStatus: PlayerFPS {
-                            idc = 27004;
-                            y = H_PART(3.3);
-                        };
-                        class Title5: Title1 {
-                            idc = -1;
-                            text = "Summary window";
-                            tooltip = "Show summary window.";
-                            y = H_PART(4.4);
-                        };
-                        class SummaryWindow: PlayerFPS {
-                            idc = 27005;
-                            y = H_PART(4.4);
-                        };
-                    };
-                };
-            };
-        };
-        class ButtonOK: ButtonOK {};
-        class ButtonCancel: ButtonCancel {};
-    };
-};
-
-
-class RscDisplayCurator {
-    class Controls {
-
-        //remove watermark
-        class Watermark {};
-
-        //summary window for diagnostics module
-        class GVAR(RscSummaryWindow): RscText {
-            idc = 95440;
-            onLoad = QUOTE(uiNamespace setVariable [ARR_2(QQGVAR(summaryWindowCtrl),_this select 0)]; (_this select 0) ctrlShow (GVAR(DiagnosticsSettings) select 4));
-            onUnload = QUOTE(uiNamespace setVariable [ARR_2(QQGVAR(summaryWindowCtrl),nil)]);
-            shadow = 2;
-            colorShadow[] = {0,0,0,1};
-            x = "safezoneX + 12.5 * (((safezoneW / safezoneH) min 1.2) / 40)";
-            y = "(1.5 * ((((safezoneW / safezoneH) min 1.2) / 1.2) / 25) + (safezoneY)) + (safezoneH - 2 * ((((safezoneW / safezoneH) min 1.2) / 1.2) / 25)) - (5 * ((((safezoneW / safezoneH) min 1.2) / 1.2) / 25))";
-            w = "safezoneW - 25 * (((safezoneW / safezoneH) min 1.2) / 40)";
-            h = "5 * ((((safezoneW / safezoneH) min 1.2) / 1.2) / 25)";
-            colorBackground[] = {0.100000, 0.100000, 0.100000, 0.500000};
-        };
-    };
-};
+#include "displays\external.hpp"
+#include "displays\moduleSupplyDrop.hpp"
+#include "displays\RscModuleDiagnostics.hpp"
+#include "displays\RscDisplayCurator.hpp"
